requirements.txt
langchain
langgraph
langchain-openai
pydantic
httpx
jinja2
python-dotenv
structlog
resend
pytest
pytest-asyncio
respx
